割り当ての探索をキャンセル可能。途中の探索結果を表示
WebWorker を使用
