+++ src/Requirements/index.ts
@@ -1,3 +1,5 @@
-export type { default, Range } from './Requirements';
-export { RequirementWithChildren, RequirementWithCourses, SelectionRequirement } from './Requirements';
 export { default as getRequirementAndDictionaryFromJSON } from './getRequirementAndDictionaryFromJSON';
+export { RequirementWithChildren, RequirementWithCourses, SelectionRequirement } from './Requirements';
+export type { default, Range } from './Requirements';
+export type { default as RequirementsJSON, RequirementWithChildrenJSON, RequirementWithCoursesJSON, SelectionRequirementJSON } from './RequirementsJSON';
+
