+++ src/V0.tsx
@@ -1,238 +1,41 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { Badge, Button, ListGroup, Modal, Spinner } from "react-bootstrap";
+// eslint-disable-next-line import/no-webpack-loader-syntax
+import V0Worker from 'worker-loader!./v0.worker';
 import Course from './Course';
-import Plan, { RegisteredCreditsCounts, RegistrationStatus } from './Plan';
-import Requirements, { Range, RequirementWithCourses } from './Requirements';
-
-function* f01(
-    requiredCreditsCount: Range,
-    acquiredCourses: readonly Course[],
-    registeredCreditsCountSum: number,
-    acquiredCreditsCountSum: number,
-    selectedCourses: readonly Course[],
-    selectedCreditsCountMin: number,
-): Generator<readonly Course[], void, undefined> {
-    if (
-        acquiredCreditsCountSum >= requiredCreditsCount.min &&
-        selectedCreditsCountMin > acquiredCreditsCountSum - requiredCreditsCount.max
-    ) {
-        yield selectedCourses;
-    }
-    if (acquiredCreditsCountSum < requiredCreditsCount.max) {
-        for (const [index, course] of acquiredCourses.entries()) {
-            const slicedCourseList = acquiredCourses.slice(index + 1);
-            const courseLists = f01(
-                requiredCreditsCount,
-                slicedCourseList,
-                registeredCreditsCountSum + course.creditsCount,
-                acquiredCreditsCountSum + course.creditsCount,
-                [...selectedCourses, course],
-                Math.min(selectedCreditsCountMin, course.creditsCount),
-            )
-            yield* courseLists;
-        }
-    }
-}
-
-function* f02(
-    requiredCreditsCount: Range,
-    registeredCourses: readonly Course[],
-    registeredCreditsCountSum: number,
-    acquiredCreditsCountSum: number,
-    selectedCourses: readonly Course[],
-    selectedCreditsCountMin: number,
-): Generator<readonly Course[], void, undefined> {
-    if (
-        registeredCreditsCountSum >= requiredCreditsCount.min &&
-        selectedCreditsCountMin > registeredCreditsCountSum - requiredCreditsCount.max
-    ) {
-        yield selectedCourses;
-    }
-    if (registeredCreditsCountSum < requiredCreditsCount.max) {
-        for (const [index, course] of registeredCourses.entries()) {
-            const slicedCourseList = registeredCourses.slice(index + 1);
-            const courseLists = f02(
-                requiredCreditsCount,
-                slicedCourseList,
-                registeredCreditsCountSum + course.creditsCount,
-                acquiredCreditsCountSum,
-                [...selectedCourses, course],
-                Math.min(selectedCreditsCountMin, course.creditsCount),
-            );
-            yield* courseLists;
-        }
-    }
-}
-
-function* f0(
-    requiredCreditsCount: Range,
-    registeredCourses: readonly Course[],
-    acquiredCourses: readonly Course[],
-    registeredCreditsCountSum: number,
-    acquiredCreditsCountSum: number,
-    selectedCourses: readonly Course[],
-    selectedCreditsCountMin: number,
-): Generator<readonly Course[], void, undefined> {
-    if (registeredCreditsCountSum < requiredCreditsCount.max) {
-        yield* f02(
-            requiredCreditsCount,
-            registeredCourses,
-            registeredCreditsCountSum,
-            acquiredCreditsCountSum,
-            selectedCourses,
-            selectedCreditsCountMin,
-        );
-        if (acquiredCreditsCountSum < requiredCreditsCount.max) {
-            for (const [index, course] of acquiredCourses.entries()) {
-                const slicedCourseList = acquiredCourses.slice(index + 1);
-                const courseLists = f0(
-                    requiredCreditsCount,
-                    registeredCourses,
-                    slicedCourseList,
-                    registeredCreditsCountSum + course.creditsCount,
-                    acquiredCreditsCountSum + course.creditsCount,
-                    [...selectedCourses, course],
-                    Math.min(selectedCreditsCountMin, course.creditsCount),
-                )
-                yield* courseLists;
-            }
-        }
-    } else {
-        yield* f01(
-            requiredCreditsCount,
-            acquiredCourses,
-            registeredCreditsCountSum,
-            acquiredCreditsCountSum,
-            selectedCourses,
-            selectedCreditsCountMin,
-        );
-    }
-}
-
-const f1 = (array: readonly {
-    requirement: RequirementWithCourses,
-    generator: Generator<readonly Course[], void, undefined>,
-}[]): {
-    requirement: RequirementWithCourses,
-    courseLists: readonly (readonly Course[])[],
-} => {
-    const array1: readonly {
-        requirement: RequirementWithCourses,
-        generator: Generator<readonly Course[], void, undefined>,
-        courseLists: (readonly Course[])[],
-    }[] = array.map(({ requirement, generator }) => ({
-        requirement, generator,
-        courseLists: [],
-    }));
-    while (true) {
-        for (const { requirement, generator, courseLists } of array1) {
-            const result = generator.next();
-            if (result.done) {
-                return { requirement, courseLists };
-            } else {
-                courseLists.push(result.value);
-            }
-        }
-    }
-}
-
-function* f2(requirements: readonly RequirementWithCourses[], plan: Plan): Generator<Plan, void, undefined> {
-    if (requirements.length === 0) {
-        yield plan;
-        return;
-    }
-
-    const t0: readonly {
-        requirement: RequirementWithCourses,
-        generator: Generator<readonly Course[], void, undefined>,
-    }[] = requirements.map(requirement => {
-        const registeredCourses = requirement.courses.filter(course => plan.courseToRequirement.get(course) === undefined && plan.courseToStatus.get(course) === RegistrationStatus.Registered);
-        const acquiredCourses = requirement.courses.filter(course => plan.courseToRequirement.get(course) === undefined && plan.courseToStatus.get(course) === RegistrationStatus.Acquired);
-        const generator = f0(
-            requirement.creditsCount,
-            registeredCourses,
-            acquiredCourses,
-            plan.requirementToOthersCount.get(requirement)?.registered ?? 0,
-            plan.requirementToOthersCount.get(requirement)?.acquired ?? 0,
-            [],
-            Infinity,
-        );
-        return { requirement, generator };
-    });
-
-    const t1 = f1(t0);
-
-    for (const courses of t1.courseLists) {
-        const plan0: Plan = {
-            ...plan,
-            courseToRequirement: new Map([
-                ...plan.courseToRequirement,
-                ...courses.map(course => [course, t1.requirement] as const),
-            ])
-        };
-        const plans = f2(requirements.filter(requirement => requirement !== t1.requirement), plan0)
-        yield* plans;
-    }
-}
-
-function* f3(requirement: Requirements, plan: Plan): Generator<readonly Plan[], void, undefined> {
-    const requirements = requirement.f(plan.selectionNameToOptionName);
-    const plan0 = { ...plan, courseToRequirement: new Map() };
-    let t0: readonly {
-        plan: Plan,
-        creditsCounts: RegisteredCreditsCounts,
-    }[] | undefined = undefined;
-    for (const plan1 of f2(requirements, plan0)) {
-        const creditsCounts = requirement.getRegisteredCreditsCounts(plan1, false);
-        if (t0 === undefined) {
-            t0 = [{
-                plan: plan1,
-                creditsCounts,
-            }];
-            yield [plan1];
-        } else {
-            let t1 = false;
-            const t2: {
-                plan: Plan,
-                creditsCounts: RegisteredCreditsCounts,
-            }[] = [];
-            for (const t3 of t0) {
-                if (!t1 && (
-                    creditsCounts.acquired > t3.creditsCounts.acquired ||
-                    creditsCounts.registered > t3.creditsCounts.registered
-                )) {
-                    t2.push({
-                        plan: plan1,
-                        creditsCounts,
-                    });
-                    t1 = true;
-                }
-                if ((
-                    t3.creditsCounts.acquired > creditsCounts.acquired ||
-                    t3.creditsCounts.registered > creditsCounts.registered
-                ) && (
-                        t3.creditsCounts.acquired === creditsCounts.acquired &&
-                        t3.creditsCounts.registered === creditsCounts.registered
-                    )) {
-                    t2.push(t3);
-                }
-            }
-            if (t1) {
-                t0 = t2;
-                yield t0.map(({ plan }) => plan);
-            }
-        }
-    }
-}
+import Plan, { fromJSON, PlanJSON, RegistrationStatus, toJSON } from './Plan';
+import Requirements, { RequirementWithCourses } from './Requirements';
 
 const V0: React.FC<{
     requirement: Requirements,
+    nameToRequirement: ReadonlyMap<string, RequirementWithCourses>,
+    codeToCourse: ReadonlyMap<string, Course>,
     plan: Plan,
     onSubmit: (plan: Plan) => void,
-}> = ({ requirement, plan, onSubmit }) => {
+}> = ({ requirement, nameToRequirement, codeToCourse, plan, onSubmit }) => {
+    const [worker, setWorker] = useState<Worker | undefined>(undefined);
     const [show, setShow] = useState(false);
     const [isLoading, setIsLoading] = useState(false);
     const [plans, setPlans] = useState<readonly Plan[] | undefined>(undefined);
+
+    const onMessage = (event: MessageEvent) => {
+        if (event.data === 'done') {
+            setIsLoading(false);
+        } else {
+            const planJSONList: readonly PlanJSON[] = event.data;
+            const plans: readonly Plan[] = planJSONList.map(planJSON => fromJSON(planJSON, { codeToCourse, nameToRequirement }));
+            setPlans(plans);
+        }
+    }
+
+    const handleHide = () => {
+        setShow(false);
+        setWorker(undefined);
+    }
+
+    useEffect(() => () => {
+        worker?.terminate();
+    }, [worker]);
 
     return (
         <>
@@ -243,16 +46,20 @@ const V0: React.FC<{
                         setShow(true);
                         setIsLoading(true);
                         setPlans(undefined);
-                        for (const p of f3(requirement, plan)) {
-                            setPlans(p);
-                        }
-                        setIsLoading(false);
+                        const worker = new V0Worker();
+                        setWorker(worker);
+                        worker.addEventListener('message', onMessage);
+                        worker.postMessage({
+                            codeToCourse,
+                            planJSON: toJSON(plan),
+                            requirementJSON: requirement.toJSON(),
+                        });
                     }
                 }
             >
                 要件を満たす割り当てを見つける
             </Button>
-            <Modal show={show} onHide={() => setShow(false)}>
+            <Modal show={show} onHide={handleHide}>
                 <Modal.Header closeButton>
                     <Modal.Title>要件を満たす割り当てを見つける</Modal.Title>
                 </Modal.Header>
@@ -267,7 +74,7 @@ const V0: React.FC<{
                             (
                                 <>
                                     <p>見つかった割り当てが以下に表示されています。適用するものを選択します。</p>
-                                    <ListGroup>
+                                    <ListGroup className={isLoading ? 'mb-3' : undefined}>
                                         {
                                             plans.map(plan1 => {
                                                 const status = requirement.getStatus(plan1);
@@ -276,7 +83,7 @@ const V0: React.FC<{
                                                     <ListGroup.Item
                                                         key={`${creditsCounts.acquired}-${creditsCounts.registered}`}
                                                         action
-                                                        onClick={() => { setShow(false); onSubmit(plan1); }}
+                                                        onClick={() => { handleHide(); onSubmit(plan1); }}
                                                     >
                                                         <div className="d-flex justify-content-between align-items-center">
                                                             <div>
@@ -304,7 +111,7 @@ const V0: React.FC<{
                         isLoading ? (
                             <>
                                 <p>
-                                    割り当てを探しています。
+                                    {plans === undefined ? '' : '単位数がより大きくなるような'}割り当てを探しています。
                                     この処理は短時間で終了しない場合があります。
                                 </p>
                                 <div className="text-center">
@@ -315,7 +122,7 @@ const V0: React.FC<{
                     }
                 </Modal.Body>
                 <Modal.Footer>
-                    <Button variant="secondary" onClick={() => setShow(false)}>キャンセル</Button>
+                    <Button variant="secondary" onClick={handleHide}>キャンセル</Button>
                 </Modal.Footer>
             </Modal>
         </>
