+++ src/react-app-env.d.ts
@@ -1,1 +1,9 @@
 /// <reference types="react-scripts" />
+
+declare module "worker-loader!*" {
+  class WebpackWorker extends Worker {
+    constructor();
+  }
+
+  export default WebpackWorker;
+}
