+++ tsconfig.json
@@ -4,7 +4,8 @@
         "lib": [
             "dom",
             "dom.iterable",
-            "esnext"
+            "esnext",
+            "WebWorker"
         ],
         "allowJs": true,
         "skipLibCheck": true,
